minor updates to help along
It was only doing one loop because of the return was nested too far. Also, I added in a bit more functionality on default values and set the projection based on the point layer selected. It still needs to to be fixed to output to a subfolder with a name specific to the transect number being run etc...